+++ lib/src/dictosaurus.dart
@@ -8,7 +8,7 @@ import 'package:dictosaurus/src/_index.dart';
 /// [Thesaurus] and [AutoCorrect] which it uses to expose the following
 /// methods:
 /// - [definitionsOf] the meaning of a term from a [DictionaryMap];
-/// - [synonymsOf] returns the synonyms of a term from a [SynonymsMap];
+/// - [synonymsOf] returns the synonyms of a term from a [Set<String>];
 /// - [suggestionsFor] returns alternative spellings for a term;
 /// - [startsWith] returns terms from a [KGramsMap] that start with a sequence
 ///   of characters; and
@@ -17,31 +17,16 @@ abstract class DictoSaurus {
   //
 
   /// Initializes a [DictoSaurus] with [autoCorrect], [thesaurus] and
-  /// [vocabulary] instances.
+  /// [dictionary] instances.
   factory DictoSaurus(
-          {required Dictionary vocabulary,
-          required Thesaurus thesaurus,
-          required AutoCorrect autoCorrect}) =>
-      _DictoSaurusImpl(autoCorrect, thesaurus, vocabulary);
+          {required Dictionary dictionary, required AutoCorrect autoCorrect}) =>
+      _DictoSaurusImpl(autoCorrect, dictionary);
 
-  /// The [DictoSaurus.inMemory] factory constructor initializes a [DictoSaurus]
-  /// with in-memory [vocabulary], [synonymsIndex] and [kGramIndex] hashmaps.
-  factory DictoSaurus.inMemory(
-          {required DictionaryMap vocabularyIndex,
-          required SynonymsMap synonymsIndex,
-          required KGramsMap kGramIndex,
-          Tokenizer? tokenizer}) =>
-      _InMemoryDictoSaurus(vocabularyIndex, synonymsIndex, kGramIndex,
-          tokenizer ?? TextTokenizer().tokenize);
-
-  /// Returns the meaning(s) of [term] and its tokenized versions.
-  ///
-  /// Returns an empty [DictionaryMap] if no meanings were found for the
-  /// [term] or any of its tokenized versions.
-  Future<DictionaryMap> definitionsOf(Term term);
+  /// Returns the meaning of [term].
+  Future<DictionaryEntry?> getEntry(Term term);
 
   /// Returns the synonyms of [term] and its tokenized versions.
-  Future<SynonymsMap> synonymsOf(String term);
+  Future<Set<String>> synonymsOf(String term);
 
   /// Returns a set of unique alternative spellings for a [term].
   ///
@@ -53,10 +38,10 @@ abstract class DictoSaurus {
 
   /// Expands [term] to an ordered list of terms.
   ///
-  /// If [term] exists in the vocabulary, the returned list will include
+  /// If [term] exists in the dictionary, the returned list will include
   /// the [term] and then its synonyms.
   ///
-  /// If [term] is not found in the vocabulary, terms with similar spellings
+  /// If [term] is not found in the dictionary, terms with similar spellings
   /// will be returned, ordered in descending order of relevance (i.e. best
   /// match first).
   ///
@@ -69,8 +54,7 @@ abstract class DictoSaurus {
 /// Sub-classes must override:
 /// - [autoCorrect] a [AutoCorrect] instance used by [suggestionsFor] and
 ///   [startsWith];
-/// - [thesaurus] a [Thesaurus] instance used by [synonymsOf]; and
-/// - [vocabulary] a [Dictionary] instance used by [definitionsOf].
+/// - [dictionary] a [Dictionary] instance used by [definitionsOf].
 ///
 abstract class DictoSaurusBase with DictoSaurusMixin {
   /// A const generative constructor for sub classes.
@@ -82,15 +66,12 @@ abstract class DictoSaurusBase with DictoSaurusMixin {
 /// [DictoSaurus.expandTerm].
 ///
 /// Implementations that mix in [DictoSaurusMixin] must override:
-/// - [vocabulary] is ;
+/// - [dictionary] is ;
 abstract class DictoSaurusMixin implements DictoSaurus {
   //
 
   /// A [Dictionary] instance used by [definitionsOf].
-  Dictionary get vocabulary;
-
-  /// A [Thesaurus] instance used by [synonymsOf].
-  Thesaurus get thesaurus;
+  Dictionary get dictionary;
 
   /// A [AutoCorrect] instance used by [suggestionsFor].
   AutoCorrect get autoCorrect;
@@ -98,13 +79,10 @@ abstract class DictoSaurusMixin implements DictoSaurus {
   @override
   Future<List<String>> expandTerm(String term, [int limit = 5]) async {
     final retVal = <Term>[];
-    final definitions = await vocabulary.definitionsOf(term);
-    if (definitions.isNotEmpty) {
+    final entry = await dictionary.getEntry(term);
+    if (entry != null) {
       retVal.add(term);
-      final synonyms = await thesaurus.synonymsOf(term);
-      for (final s in synonyms.values) {
-        retVal.addAll(s);
-      }
+      retVal.addAll(entry.allSynonyms);
     } else {
       final suggestions = await autoCorrect.suggestionsFor(term, limit);
       for (final s in suggestions.values) {
@@ -115,11 +93,11 @@ abstract class DictoSaurusMixin implements DictoSaurus {
   }
 
   @override
-  Future<DictionaryMap> definitionsOf(Term term) =>
-      vocabulary.definitionsOf(term);
+  Future<DictionaryEntry?> getEntry(Term term) => dictionary.getEntry(term);
 
   @override
-  Future<SynonymsMap> synonymsOf(String term) => thesaurus.synonymsOf(term);
+  Future<Set<String>> synonymsOf(String term) async =>
+      (await getEntry(term))?.allSynonyms ?? {};
 
   @override
   Future<List<String>> startsWith(String chars) =>
@@ -131,29 +109,8 @@ abstract class DictoSaurusMixin implements DictoSaurus {
       autoCorrect.suggestionsFor(term, limit);
 }
 
-/// Implementation of [DictoSaurus] with in-memory [DictionaryMap],
-/// [SynonymsMap] and [KGramsMap] hashmaps.
-class _InMemoryDictoSaurus with DictoSaurusMixin {
-  /// Instantiates a [_InMemoryDictoSaurus] with in-memory [vocabulary],
-  /// [synonymsIndex] and [kGramIndex] hashmaps.
-  _InMemoryDictoSaurus(DictionaryMap vocabularyIndex, SynonymsMap synonymsIndex,
-      KGramsMap kGramIndex, Tokenizer tokenizer)
-      : vocabulary = Dictionary.inMemory(vocabularyIndex, tokenizer: tokenizer),
-        thesaurus = Thesaurus.inMemory(synonymsIndex, tokenizer: tokenizer),
-        autoCorrect = AutoCorrect.inMemory(kGramIndex, tokenizer: tokenizer);
-
-  @override
-  final AutoCorrect autoCorrect;
-
-  @override
-  late Thesaurus thesaurus;
-
-  @override
-  final Dictionary vocabulary;
-}
-
 /// A [DictoSaurus] with final [autoCorrect], [thesaurus] and
-/// [vocabulary] fields and a generative constructor.
+/// [dictionary] fields and a generative constructor.
 class _DictoSaurusImpl extends DictoSaurusBase {
   //
 
@@ -161,12 +118,9 @@ class _DictoSaurusImpl extends DictoSaurusBase {
   final AutoCorrect autoCorrect;
 
   @override
-  final Thesaurus thesaurus;
-
-  @override
-  final Dictionary vocabulary;
+  final Dictionary dictionary;
 
   /// Initializes a [DictoSaurus] with [autoCorrect], [thesaurus] and
-  /// [vocabulary] instances.
-  const _DictoSaurusImpl(this.autoCorrect, this.thesaurus, this.vocabulary);
+  /// [dictionary] instances.
+  const _DictoSaurusImpl(this.autoCorrect, this.dictionary);
 }